CORE-4474 Refactored IplantAnnouncementConfig.
Refactored IplantAnnouncementConfig, IplantAnnouncement, and IplantAnnouncer so that the config can be constructed with the message that is returned in a Widget for use by the Announcement, rather than passing the message and config separately in an IplantAnnouncer::schedule call.
Updated ErrorAnnouncementConfig styling to include an error icon before the message.
Added a SuccessAnnouncementConfig that includes a success icon before the message.