CORE-4600 Initial data service call caching.
DiskResourceServiceFacadeImpl now extends a Folder TreeStore, so that it can cache folders and files that are fetched from service calls, then return them for subsequent service calls.
Refactored the 2 DiskResourceServiceFacade#getFolderContents methods into getSubFolders and getFolderContents methods, which use AsyncCallbackConverters and accept AsyncCallbacks with stronger return types.
So far, these 2 methods and getRootFolders are the only calls caching service responses and using cached results, which leaves the other actions (such as moving and deleting) in a broken state.
Added a RootFolders#setRoots method.