CORE-4685 Refactored Max/Min Number Validators.
Renamed the Min/MaxDoubleValidator to DoubleAbove/BelowValidator.
Renamed the Min/MaxIntegerValidator to IntAbove/BelowValidator.
Changed the logic of these Validators to validate their given values exclusive of their limits. For example, IntAboveValidator will validate that the given value is greater than its limit, instead of greater than or equal to its limit, which is the default behavior of GXT's MinNumberValidator.
Updated the validation message in NumberRangeValidator.